CIV-17524 Add/fix unit tests
